requirements.txt
torch
torchvision
ultralytics
opencv-python
pillow
numpy